backend/pyproject.toml
[tool.poetry]
name = "auditronai-backend"
version = "0.1.0"
description = "Backend pour l'application AuditronAI"
authors = ["Your Name <your.email@example.com>"]

[tool.poetry.dependencies]
python = "^3.9"
fastapi = "^0.104.1"
uvicorn = "^0.24.0"
python-multipart = "^0.0.6"
aiofiles = "^23.2.1"
aiohttp = "^3.9.1"
pydantic = {extras = ["email"], version = "^2.5.2"}
sqlalchemy = "^2.0.23"
alembic = "^1.13.0"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
python-dotenv = "^1.0.0"
asyncpg = "^0.29.0"
aiosqlite = "^0.19.0"
pytest = "^7.4.3"
pytest-asyncio = "^0.21.1"
httpx = "^0.25.2"

[tool.poetry.dev-dependencies]
black = "^23.11.0"
isort = "^5.12.0"
flake8 = "^6.1.0"
mypy = "^1.7.1"
pytest-cov = "^4.1.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88
target-version = ['py39']
include = '\.pyi?$'

[tool.isort]
profile = "black"
multi_line_output = 3
